Fix url scheme mismatch in SSO login and out
When the login page on a network site is https, the SSO script will be
blocked because of protocol mismatch when it tries to load each site
over the http url which is generated by `esc_url( $site['domain'] )`.

This fixes that issue by setting the URL scheme to the same as the
current site's login scheme (which should really be https on any modern
sites).