Prevent duplicate inserts into domain mapping logins table
When using this plugin along with the WP MU Domain Mapping plugin, it's
possible for the same domain to be returned twice in the response for
WP_Multisite_SSO::get_network_sites() if a domain is used as an alias of
another blog.

This commit removes duplicate domains from the return value here, so that we
don't try to log into the same domain twice. (Logging into the same
domain twice causes issues duplicate insert issues, because the blog ID
of a domain is used along with the current timestamp to build the insert
key in the wp_domain_mapping_logins table.)